extra listContent and test it directly
